Refactor SpectraGrouping to Spectra for improved structure
This commit refactors the SpectraGrouping functionality in the CompMs application by renaming namespaces and classes to `Spectra`. Key changes include updates to `MainWindow.xaml.cs`, `SpectraGroupingView.xaml`, `SpectraGroupingView.xaml.cs`, and `LcmsAlignmentViewModel.cs` to reflect the new organization.

The ViewModel now utilizes a `SpectraGroupingModel`, enhancing data management and responsiveness. Outdated properties and classes, such as `MoleculeGroupModel` and `ProductIonAbundanceModel`, have been removed in favor of a more streamlined approach using reactive properties for better data binding.

These changes improve the maintainability and clarity of the codebase, facilitating easier management of Spectra-related functionalities.
