+++ src/MSDIAL5/MsdialGuiApp/View/Core/MainWindow.xaml.cs
@@ -8,7 +8,7 @@ using CompMs.App.Msdial.View.MsResult;
 using CompMs.App.Msdial.View.PeakCuration;
 using CompMs.App.Msdial.View.Search;
 using CompMs.App.Msdial.View.Setting;
-using CompMs.App.Msdial.View.SpectraGrouping;
+using CompMs.App.Msdial.View.Spectra;
 using CompMs.App.Msdial.View.Statistics;
 using CompMs.App.Msdial.View.Table;
 using CompMs.App.Msdial.ViewModel.Chart;
@@ -20,7 +20,7 @@ using CompMs.App.Msdial.ViewModel.PeakCuration;
 using CompMs.App.Msdial.ViewModel.Search;
 using CompMs.App.Msdial.ViewModel.Service;
 using CompMs.App.Msdial.ViewModel.Setting;
-using CompMs.App.Msdial.ViewModel.SpectraGrouping;
+using CompMs.App.Msdial.ViewModel.Spectra;
 using CompMs.App.Msdial.ViewModel.Statistics;
 using CompMs.App.Msdial.ViewModel.Table;
 using CompMs.CommonMVVM.Common;
