+++ src/MSDIAL5/MsdialGuiApp/View/Spectra/SpectraGroupingView.xaml.cs
@@ -1,6 +1,6 @@
 ﻿using System.Windows.Controls;
 
-namespace CompMs.App.Msdial.View.SpectraGrouping;
+namespace CompMs.App.Msdial.View.Spectra;
 
 public partial class SpectraGroupingView : UserControl
 {
