+++ src/MSDIAL5/MsdialGuiApp/ViewModel/Lcms/LcmsAlignmentViewModel.cs
@@ -7,7 +7,7 @@ using CompMs.App.Msdial.ViewModel.Information;
 using CompMs.App.Msdial.ViewModel.Search;
 using CompMs.App.Msdial.ViewModel.Service;
 using CompMs.App.Msdial.ViewModel.Setting;
-using CompMs.App.Msdial.ViewModel.SpectraGrouping;
+using CompMs.App.Msdial.ViewModel.Spectra;
 using CompMs.App.Msdial.ViewModel.Statistics;
 using CompMs.App.Msdial.ViewModel.Table;
 using CompMs.Common.Enum;
@@ -49,7 +49,7 @@ namespace CompMs.App.Msdial.ViewModel.Lcms
 
             var spectraSimilarityMapViewModel = new SpectraSimilarityMapViewModel(model.SpectraSimilarityMapModel).AddTo(Disposables);
             var spectraSmilarityMapCommand = new ReactiveCommand().WithSubscribe(() => broker.Publish(spectraSimilarityMapViewModel)).AddTo(Disposables);
-            var spectraGroupingViewModel = new SpectraGroupingViewModel().AddTo(Disposables);
+            var spectraGroupingViewModel = new SpectraGroupingViewModel(model.SpectraGroupingModel).AddTo(Disposables);
             var spectraGroupingCommand = new ReactiveCommand().WithSubscribe(() => broker.Publish(spectraGroupingViewModel)).AddTo(Disposables);
 
             var (peakPlotAction, peakPlotFocused) = focusControlManager.Request();
