+++ src/MSDIAL5/MsdialGuiApp/ViewModel/Spectra/SpectraGroupingViewModel.cs
@@ -1,0 +1,35 @@
+using CompMs.App.Msdial.Model.Chart;
+using CompMs.App.Msdial.Model.DataObj;
+using CompMs.App.Msdial.Model.Spectra; 
+using CompMs.CommonMVVM;
+using Reactive.Bindings;
+using Reactive.Bindings.Extensions;
+using System.Collections.ObjectModel;
+
+namespace CompMs.App.Msdial.ViewModel.Spectra; 
+
+/// <summary>
+/// ViewModel for Spectra Grouping View.
+/// </summary>
+public class SpectraGroupingViewModel : ViewModelBase {
+    private readonly SpectraGroupingModel _model;
+
+    public SpectraGroupingViewModel(SpectraGroupingModel model) {
+        _model = model;
+        MoleculeGroups = new ReadOnlyObservableCollection<MoleculeGroupModel>(model.MoleculeGroups);
+        SelectedMoleculeGroup = model.ToReactivePropertySlimAsSynchronized(m => m.SelectedMoleculeGroup).AddTo(Disposables);
+        ProductIonAbundances = model.ObserveProperty(m => m.ProductIonAbundances).ToReadOnlyReactivePropertySlim().AddTo(Disposables);
+        TheoreticalSpectra = model.ObserveProperty(m => m.TheoreticalSpectra).ToReadOnlyReactivePropertySlim().AddTo(Disposables);
+        SelectedSample = model.ToReactivePropertySlimAsSynchronized(m => m.SelectedSample).AddTo(Disposables);
+        MeasuredSpectra = model.ObserveProperty(m => m.MeasuredSpectra).ToReadOnlyReactivePropertySlim().AddTo(Disposables);
+    }
+
+    public ReadOnlyObservableCollection<MoleculeGroupModel> MoleculeGroups { get; }
+    public ReactivePropertySlim<MoleculeGroupModel?> SelectedMoleculeGroup { get; }
+
+    public ReadOnlyReactivePropertySlim<ProductIonAbundanceModel[][]?> ProductIonAbundances { get; }
+    public ReadOnlyReactivePropertySlim<MsSpectrum?> TheoreticalSpectra { get; }
+    public AnalysisFileBeanModel[] Samples => _model.Samples;
+    public ReactivePropertySlim<AnalysisFileBeanModel?> SelectedSample { get; }
+    public ReadOnlyReactivePropertySlim<MsSpectrum?> MeasuredSpectra { get; }
+}
